Move two functions into separate file
