+++ CMakeLists.txt
@@ -45,6 +45,7 @@ set(discord-screenaudio_SRC
   src/log.cpp
   src/userscript.cpp
   src/centralwidget.cpp
+  src/utils.cpp
   resources.qrc
 )
